+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Align field names with the EURING manual (#46).
+- Add `convert --file` support for converting record files.
 
 ## 25.2 (2025-12-30)